macos: Ghostty.Config to store all config-related operations
